refactor(useMediaQuery): To be stored inside ThemeProvider context
