+++ src/hooks/useMediaQuery/index.js
@@ -1,85 +1,18 @@
 // @flow
 import * as React from "react";
 
-import useTheme from "../useTheme";
-import { getBreakpointWidth } from "../../utils/mediaQuery/index";
+import QueryContext from "../../ThemeProvider/QueryContext";
+import useMediaQueryContext from "../../ThemeProvider/QueryContext/useMediaQueryContext";
 
-import type { UseMediaQuery } from ".";
-
-const createBreakpoints = theme => ({
-  isDesktop: getBreakpointWidth("desktop", theme),
-  isLargeDesktop: getBreakpointWidth("largeDesktop", theme),
-  isTablet: getBreakpointWidth("tablet", theme),
-  isLargeMobile: getBreakpointWidth("largeMobile", theme),
-  isMediumMobile: getBreakpointWidth("mediumMobile", theme),
-});
-
-const createMediaQueryList = breakpoints => {
-  if (typeof window !== "undefined") {
-    /* $FlowFixMe(>=0.115.0) This comment suppresses an error found when upgrading Flow to
-     * v0.115.0. To view the error, delete this comment and run Flow. */
-    return Object.keys(breakpoints).map(q => ({ [q]: window.matchMedia(breakpoints[q]) }));
-  }
-  return [];
-};
+import type { UseMediaQuery } from "./index";
 
 const useMediaQuery: UseMediaQuery = () => {
-  const theme = useTheme();
-  const breakpoints = React.useMemo(() => createBreakpoints(theme), [theme]);
-  const mediaQueryLists = React.useMemo(
-    () =>
-      Object.assign(
-        {
-          isDesktop: {},
-          isLargeDesktop: {},
-          isTablet: {},
-          isLargeMobile: {},
-          isMediumMobile: {},
-        },
-        ...createMediaQueryList(breakpoints),
-      ),
-    [breakpoints],
-  );
-  const timeoutRef = React.useRef(null);
-  const getValue = React.useCallback(() => {
-    return Object.assign(
-      {
-        isDesktop: false,
-        isLargeDesktop: false,
-        isTablet: false,
-        isLargeMobile: false,
-        isMediumMobile: false,
-      },
-      /* $FlowFixMe(>=0.115.0) This comment suppresses an error found when upgrading Flow to
-       * v0.115.0. To view the error, delete this comment and run Flow. */
-      ...Object.keys(mediaQueryLists).map(q => ({ [q]: mediaQueryLists[q]?.matches })),
-    );
-  }, [mediaQueryLists]);
-
-  const [value, setValue] = React.useState(getValue);
-
-  React.useEffect(() => {
-    setValue(getValue);
-    const handler = () => {
-      if (typeof setTimeout === "function" && typeof clearTimeout === "function") {
-        clearTimeout(timeoutRef.current);
-        timeoutRef.current = setTimeout(() => {
-          timeoutRef.current = null;
-          setValue(getValue);
-        }, 150);
-      }
-    };
-    Object.keys(mediaQueryLists).forEach(mql => {
-      mediaQueryLists[mql].addListener(handler);
-    });
-    return () => {
-      Object.keys(mediaQueryLists).forEach(mql => mediaQueryLists[mql].addListener(handler));
-      if (timeoutRef.current !== null && typeof clearTimeout === "function") {
-        clearTimeout(timeoutRef.current);
-      }
-    };
-  }, [getValue, mediaQueryLists]);
-  return value;
+  const media = React.useContext(QueryContext);
+  if (media != null && Object.values(media).some(v => v != null)) {
+    return media;
+  }
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  return useMediaQueryContext();
 };
 
 export default useMediaQuery;
